+++ src/chatdDb.h
@@ -108,7 +108,7 @@ public:
             stmt.blobCol(4, buf);
 #ifndef NDEBUG
             auto idx = stmt.intCol(5);
-            assert(idx == mMessages.lownum()-1-messages.size());
+            assert(idx == mMessages.lownum()-1-(int)messages.size());
 #endif
             auto msg = new chatd::Message(msgid, userid, ts, std::move(buf),
                 stmt.intCol(7), (chatd::Message::Type)stmt.intCol(3));
